Create index.html.tera
So before when the check button was pressed the only thing that popped up on the webpage was the output thus every time the user had to enter the new value he had to go back to the previous page so as to re-enter the currency which was a bad approach. In order to overcome this challenge, I added an index.tera.html file which had the same structure as that of the index.html file but it had an extra field which was the output textual currency. So previously my return type for the sink1() function was "String" which was now changed to "Template" and this template was returned.